requirements.txt
langchain
langchain-community
PyPDF2
streamlit
openai
faiss-cpu
pandas
numpy
python-dotenv
